upstream: b=main,r=f12b3e997d54dca2f933c5991c1906e7cc5e2509,t=2023-11-24-1342-52271

+++ components/nexus-repository-content/src/main/java/org/sonatype/nexus/repository/content/browse/store/BrowseNodeData.java
@@ -36,7 +36,7 @@ import static org.sonatype.nexus.repository.content.store.InternalIds.toExternal
 public class BrowseNodeData
     implements BrowseNode
 {
-  Integer nodeId; // NOSONAR: internal id
+  Long nodeId; // NOSONAR: internal id
 
   int repositoryId; // NOSONAR: internal repository id
 
@@ -44,7 +44,7 @@ public class BrowseNodeData
 
   private String displayName;
 
-  int parentId; // NOSONAR: internal id
+  long parentId; // NOSONAR: internal id
 
   private boolean leaf;
   
@@ -95,7 +95,7 @@ public class BrowseNodeData
   /**
    * Sets the internal node id.
    */
-  public void setNodeId(final int nodeId) {
+  public void setNodeId(final long nodeId) {
     this.nodeId = nodeId;
   }
 
@@ -123,7 +123,7 @@ public class BrowseNodeData
   /**
    * Sets the internal parent node id.
    */
-  public void setParentId(final int parentId) {
+  public void setParentId(final long parentId) {
     this.parentId = parentId;
   }
